flaskPostgresProjectTemplate/requirements.txt
Flask
psycopg2-binary
sqlparse